+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Styling improvements to webhook and pubsub table in Console.
+- Gateway location is updated even if no antenna locations had been previously set.
 
 ### Deprecated
 
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - End device claim display bug when claim dates not set.
 - DeviceModeInd handling for LoRaWAN 1.1 devices.
+- Do not perform unnecessary gateway location updates.
 
 ### Security